Make instruction more idiot proof
The instructions say 

```
cd spark
MASTER=yarn-cluster bin/run-example SparkPi
```

assuming that you are already in the `/hathi-client` folder, which is not the case if you followed the instructions before without modifications. Seems like this used to be true, because there is the following commented out snipped left in the file:

```
cd hathi-client
bin/get.sh spark
```

But as mentioned, it has been commented out. Anyway this is easy to fix, if you just use the full path.